+++ tests/testthat/helper-predictor.R
@@ -1,6 +1,6 @@
 
 f = function(newdata, multi = FALSE) {
-  pred = unlist(newdata[, 1] + newdata[, 2] + 100 * (newdata[3] == "a")) / (155)
+  pred = unlist(newdata[, "a"] + newdata[, "b"] + 100 * (newdata[,"c"] == "a")) / (155)
   dat = data.frame(pred = pred)
   colnames(dat) = "pred"
   if (multi) dat$pred2 = 1 - dat$pred